requirements.txt
acres==0.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
av==14.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.1.0
ctranslate2==4.5.0
dataclasses-json==0.6.7
distro==1.9.0
etelemetry==0.3.1
fastapi==0.115.6
faster-whisper==1.1.0
ffmpy==0.4.0
filelock==3.16.1
fitz==0.0.1.dev2
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
gradio==5.8.0
gradio_client==1.5.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
importlib_resources==6.4.5
isodate==0.6.1
Jinja2==3.1.4
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.22
langchain-text-splitters==0.3.2
langsmith==0.1.147
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.1
numpy==1.26.4
onnxruntime==1.20.1
openai==1.57.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.0.0
propcache==0.2.1
protobuf==5.29.1
prov==2.0.1
puremagic==1.28
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydot==3.0.3
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.25.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
ruff==0.8.2
safehttpx==0.1.6
safetensors==0.4.5
scipy==1.14.1
semantic-version==2.10.0
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.21.0
tomlkit==0.13.2
tqdm==4.67.1
traits==6.4.3
transformers==4.47.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==14.1
yarl==1.18.3
